langgraph-standard/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
tavily-python>=0.5.0
amp-instrumentation>=0.1.2
